Create a new font config for each document
