+++ weasyprint/css/__init__.py
@@ -30,7 +30,7 @@ from . import properties
 from . import computed_values
 from .descriptors import preprocess_descriptors
 from .validation import preprocess_declarations
-from ..text import add_font_face
+from ..fonts import add_font_face
 from ..urls import (element_base_url, get_url_attribute, url_join,
                     URLFetchingError)
 from ..logger import LOGGER
